+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Added:
 
 Changed:
 
-- `log` file is optional for `ONTraC_analysis`
+- Make `log` file optional for `ONTraC_analysis`
 
 Fixed: